Add shuffle scaffolding and wire it up with the partition processor
This commit introduces the shuffle co-routine and wires it up with the
partition processor and the network component.

The shuffle is dynamically registered at the network when a partition
processor becomes leader. Besides the network input/output channels,
it is started with an outbox reader, a command sink which is connected
to the partition processor to send outbox truncations and a hint receiver
which allows the partition processor to notify the shuffle about newly
available messages in the outbox.

This fixes #105.