Split ActuatorStream from LeadershipState
By splitting the ActuatorStream form the LeadershipState, we don't have to create
an ActuatorStream every iteration of the partition processor loop. This might make
the loop a bit more efficient.

Moreover, this PR moves the polling of the shuffle JoinHandle out of the actuator
stream into a separate method which separates concerns better.