Allow PartitionProcessor to generate ServiceInvocations from Background/Invoke journal entries
In order to allow the partition processor to generate ServiceInvocations, it first needs
to extract the service key and generate an invocation id from an Background/InvokeEntry before
it is proposed. The service key and the invocation id will be attached to the journal entry
so that it can be used by the state machine when applying the command. That way, all replicas
will see the same information.

In order to make the journal entries extendable, this commit generified the RawEntry so that it
now accepts different header types. That way we can convert a RawEntryHeader into an
EnrichedEntryHeader carrying additional information.

This fixes #148.