Implement memory limits for the timer service
This commit implements memory limits for each individual timer service. It
enforces that every partition leader only keeps the earliest #n timers in
memory. Whenever those timers have fired, the timer service will reload remaining
timers from the storage via the TimerReader.

An important contract between the timer::Service and the TimerReader is that the
timers have a Timer::TimerKey which orders the timer according to their firing time
and how the timers are read/loaded by the TimerReader. If the TimerReader does not
load the timers in this order, then the timer::Service might not process all timers
since some of them might get silently dropped.

This commit introduces a new dependency priority-queue which is licensed under
the Mozilla Public License 2.0 which is a weak copy-left license. This should be
ok in our case since we are not modifying the code of this dependency and the MPL-2.0
allows to distribute statically linked code under a different license if not modified.

This fixes #207.