Align TimerValue ordering with RocksDB key layout
We use the TimerKey to read the timers in an absolute order. The timer service
relies on this order in order to process each timer exactly once. That is the
reason why the ordering of the TimerValue and how the TimerKey is laid out in
RocksDB need to be exactly the same.