Track max fired timer in timer logic
In order to avoid re-reading fired timers from storage, the timer logic
needs to track what was the maximum fired timer so far. This value will
be the exclusive start for the next scan operation to read the next batch
of timers from RocksDB.

Before, it could happen that after a batch was finished an older timer arrives.
Once this timer is processed (immediately because it is older than the current
time), the system would read the next timers based on the last batch. Since the
timer was older, it could mean that some later but already processed and not yet
deleted timers would be read from RocksDB.