Deduplicate timers
There is a potential race condition between the partition processor
notifying the timer service about new timers that are already stored
in RocksDB and the timer service reading timers directly from RocksDB.
If the timer service reads the next timers from RocksDB just after the
partition processor stored a new timer, it can happen that this timer
is being processed before the notification via the input channel arrives.
In this case, we have to filter out the notification in order to avoid
firing the timer twice.

This commit solves the problem by introducing a timer sequence number
which is assigned for every timer by the partition processor. This sequence
number is being stored in RocksDB as well. That way, the timer service
knows whether it has already seen a given timer for which it has received
a notification via the input channel or not by comparing the sequence number
with the maximal seen sequence number so far.

This fixes #382.