Introduce Kill command to manually abort an invocation
The kill command fails an invocation with the error code 10 which
represents the aborted grpc status code. Moreover, it sends an abort
message to the invoker to stop executing this invocation.

If the specified service invocation is not executing (invoked or
suspended), then the command will be ignored.

This fixes #406.