Run TimerService as part of the partition processor
By moving the TimerService into the partition processor, we can avoid the
race condition between receiving input from the channel and reading timers
from RocksDB because the PartitionProcessor is now in charge or polling the
future. This means that when new timers are being added to the TimerService
via addTimer, it is guaranteed that these timer have not been read from RocksDB
yet.

This fixes #427.