Exclude tools packages from cargo default members
In order to not build the tools packages, they are not part of the default
members. Moreover, this commit introduces a new just command build-tools which
builds the tools.

We use --workspace for the just clippy command in order to include all tools
packages in the clippy phase (they are being compiled).

This fixes #520.