Use a single channel for sending inputs to the Invoker
This commit unifies the three input channels of the invoker into a single
input channel. This change has the advantage that input messages over this
single channel are now properly ordered. Hence, this commit solves the
race condition which caused inovcation/resumptions from being dropped
because they were sent over a different input channel than the registration
of a partition.

This fixes #529.