[NBIS] Cleanup unused code
[RemoteContext] Use journal table to store journal entries.

[PP] Support the new NBISEffects. In order to do that, I reused the same AppendJournalEntry effect used by StateMachine::handle_journal_entry. While working on that, I removed the distinction AppendJournalEntry/AppendAwakeableJournalEntry/AppendJournalEntryAndAck and splitted the InvokerEffect::StorageAck to a separate Effect.

[DBIS] Add support for reading/writing journal

Implement mechanism to deliver journal completions to new invocations.

Split the JournalMetadata type. Now the type used by invoker_api::JournalReader is different from the stored type, as both endpoint_id and method don't make sense in the general case of the journal, and they're not strictly related to the journal itself, but rather to the invocation.