[CLI] Listing services and console primitives
## Summary
This introduces the console I/O primitives that should be used for future CLI work. The set of macros in `ui/console.rs`
replace the standard `print*!()` family of macros to avoid panics on broken pipes. Additionally, it has a couple of convenient
macros and types that provides consistent styling of the console output. The stylesheet is defined at `ui/stylesheet.rs`.

This also introduces a client component for the meta HTTP service. The client currently lives within the `restate-cli` crate but
it might its own crate if needed in the future. In order to reduce code duplication, the client reuses the API model types from
`restate-meta` and `restate-schema-api`.

Temporarily, the needed API model types from `restate-meta` were made public but this will change in the next PR, which extracts
the model types into a separate crate and make those types available to all.

## Commands introduced:
`restate services list` (or for short) `restate svc ls`
Prints the list of services in a compact output by default, more information can be printed by adding `--extra` and table borders
can be controlled by `--table-style=borders` as a global UI control that will be respected by future commands.

```
     NAME                                     REV  FLAVOR    ENDPOINT
 🌎  io.shoppingcart.PaymentGateway           1              HTTP 2
 🌎  io.shoppingcart.InventoryRestockService  1              HTTP 2
 🌎  io.shoppingcart.CheckoutFlowService      1              HTTP 2
 🌎  io.shoppingcart.EmailSender              1    ⬅️ 🚶🚶🚶  HTTP 2
 🌎  io.shoppingcart.ShipmentService          1              HTTP 2
 🌎  io.shoppingcart.UserProfileService       1    ⬅️ 🚶🚶🚶  HTTP 2
 🌎  io.shoppingcart.ProductService           1    ⬅️ 🚶🚶🚶  HTTP 2
 🌎  io.shoppingcart.ShoppingCartService      1    ⬅️ 🚶🚶🚶  HTTP 2
```