[storage] Enable the strictest sync mode for RocksDB's WAL
This commit enables sync mode = true for RocksDB WAL, when committing a WriteBatch.
This option isn't enabled by default.
This would increase the durability level to survive a whole machine
crush, and with the expense of (potentially) higher commit latency.
Keep in mind that the WAL writes are temporarily enabled for the main
store, up to the point that our distributed log will be responsible for
durability.