Separate connect and request timeout; only apply default request timeout for Metas client
Since DataFusion queries may take much longer and are hard to predict, we don't apply a request timeout by default to the DataFusion HTTP client. The same command line switch controls the request timeout for both types of request, but a default is only silently applied for "known fast" operations.