Add optional Bearer Token auth and a default HTTP request timeout (#1027)
* Add a default timeout to the CLI's HTTP clients

* Add support for bearer token auth via environment variable

* Separate connect and request timeout; only apply default request timeout for Metas client

Since DataFusion queries may take much longer and are hard to predict, we don't apply a request timeout by default to the DataFusion HTTP client. The same command line switch controls the request timeout for both types of request, but a default is only silently applied for "known fast" operations.