Use restate_storage_api::TimerKey in TimerValue
This commit changes the TimerValue to use the restate_storage_api::TimerKey
as the key. This implies that it also relies on the ordering of the TimerKey
for the in-memory representation. We ensure that the TimerKey's in-memory
ordering is the same as its binary order by adding a randomized test to
catch future regressions.

This fixes #394.