Allow timers to be removed from running TimerService
This commit adds timer::Service::remove_timer which allows to remove a
timer from the TimerService while it is running. Moreover, when cancelling
a timer, we also issue a timer deletion from RocksDB and removal from the
TimerService.

If the TimerService is currently loading new timers, we need to remember the
removed timer for the time of loading new timers to filter them out. After
leaving this state, we can forget the removed timers.

If the TimerSerivce is currently processing timers, we can directly remove
timers from the internal timer queue. If the removed timer is the last timer
in the current batch, then we need to update the batch end to reflect the
new state.

This fixes #1029.