Introduce max batch size and duration for PartitionProcessor
This commit introduces an upper limit for the size of command batches
that the PartitionProcessor can create and the maximum duration the
batching can take before checking again the actuators for signals. This
prevents the situation where the PartitionProcessor is starved because
of too many commands being applied.

This fixes #1051.