Add storage format version to RocksDBStorage
This commit introduces the notion of a storage format version for
RocksDBStorage. The storage format version is stored in the default
column family under "internal_storage_format_version" and represents
the storage format of all state. When opening RocksDBStorage, it
will check whether the db is empty. If the db is empty, it will write
the current storage format version. If not, then it will check
whether the stored storage format version is compatible with the
current storage format version.

This fixes #1087.