Add version information to FileMetaStorage
This commit lets the FileMetaStorage implementation write version
information to a separate file root_directory/.meta_format_version.
It contains the json encoded version. When starting, the
FileMetaStorage reads this file and checks whether it is compatible
with the current storage formate version. In case that the storage
directory is empty, it will write a new .meta_format_version file.

This fixes #1088.