Let EffectInterpreter pop from the inbox
This commit changes how the PP state machine pops from the inbox. Instead
of letting the CommandInterpreter do it, now the EffectInterpreter is
responsible for popping the next inbox entry from the inbox. That way, we
enable that the EffectInterpreter can pop multiple inbox entries from the
inbox if needed.