Let PartitionProcessor apply Command::ExternalStateMutation commands
This commit adds support for applying Command::ExternalStateMutation commands
in the PP's state machine. The way it works is the following: Upon receiving
the command, the CommandInterpreter checks whether the service instance is
currently locked. If so, then the ExternalStateMutation is enqueued into the
inbox. If not, then the ExternalStateMutation is directly applied by the
EffectInterpreter. Upon completing an invocation, the EffectInterpreter will
now drain the inbox until it finds the next ServiceInvocation. All encountered
ExternalStateMutations will be applied when searching for the next
ServiceInvocation.

The inbox now needs to be able to store ExternalStateMutations and
ServiceInvocations. That's why this commit updates the InboxEntry type to
support both variants. The underlying storage Protobufs of the InboxEntry
were changed in a backward compatible way by introducing a new field called
entry and keeping the service_invocation field.

The storage-query-datafusion inbox table has not been update wrt to the new
inbox entry types. Instead, all state mutations are being filtered out when
querying the inbox. This is something that needs to be done as a follow-up.

This fixes #1099.