New versioned deployment ID scheme
**Note:** This is a _big_ refactoring that is hard to split, but many changes are mechanical. I'll highlight the key changes in the description to make it easier to review.


This change introduces new scheme for deployment IDs, the scheme follows the recently introduced `ResourceId` trait that aims to unify
the user-facing ID scheme in a versioned and copy-paste friendly format.

An example of the new IDs: `dp_134xwosaM2PKrh2dZTMbotj`. The prefix `dp_` is short for `deployment`.
The first character after the separator is the ID version scheme. The ID is lexographically sortable
which makes it easy to rough-order deployments by recency.

The most notable change in this PR is in `compute_new_deployment` function. The change introduces the ability
for the caller to specify a deployment-id for the new deployment, or leave it as None for an auto-generated one.

This is currently only being leveraged in tests but it can be used in future user-facing features that require editing existing deployments (adding new endpoints, etc.)
The logic is described in the code comment but the summary is:
> If the ID is set on request, we assume that the user wants to create a deployment with this ID, or if the deployment exists, the user _only_ accepts to overwrite this deployment **if** existing deployment matches with this ID.


Test Plan:
- Tests have been updated to validate the new logic and ID scheme
- System was tested manually with the CLI for a variety of cases (register, list, invocation, etc.)
- Verification/e2e tests