Subscription ID to follow resource ID scheme
This introduces a formal format for subscription IDs and moves everything to use the `SubscriptionId` newtype (implements ResourceId)

An example subscription id: `sub_15VqmTOnXH3Vv2pl5HOG7Ua` which follows the same scheme.

The most notable change is that we now don't accept `id` from the user when creating subscriptions.

Closes #1117