Move emission of message size warning and limit into DecoderState
By moving the warning and limit of the message size into the DecoderState,
we avoid the repeated emission of the warning in case that a large message
is read. Only when receiving the message header which specifies the needed
memory we emit it. The same applies to enforcing the message limit.

This fixes #1128.