Bootstrap cluster controller crate and wire it up with node
This commit introduces thes cluster controller crate. It
contains the cluster controller which currently accepts attachment
calls.

The node crate has been updated to attach itself to the
cluster controller before it starts the worker.

Whether the process should run the cluster controller is configurable
via the cluster_controller_endpoint config option which takes either
"local" or the address to the remote cluster_controller. In the former
case, the process will start a cluster controller locally.