Generational NodeID
This introduces generational node ids. A `NodeId` type can either represent a plain id, or an id with generation.
Nodes with stable ID will get an increasing generation on startup, this will be used to fence off work meant for
older instances of the node (e.g. responses for an older ingress instance), or to ensure we are sending messages
to the correct instance at send time.