Make the worker role optional
This commit changes the Node implementation to make the worker
role optional. This also includes the node-ctrl service which
can run a worker grpc service if the worker role is configured.
If a node does not run the cluster controller role locally, then
one needs to configure it with a cluster_controller_address.