Move meta and admin component to cluster controller role
This commit decouples the meta and admin components from the worker and
moves them over to the cluster controller role. The way the worker learns
about schema information changes is by periodically asking the schema
service for updates. At the moment, every schema information fetch request
will send the whole list of schema update commands over.

Currently, it is not possible to query the storage via the Admin service.
This will be fixed with a follow-up commit.