TaskCenter all the things
This migrates _most_ of the long-running tasks to task center, and we get to harvest the fruit of this investment.
Note that this is still not 100% of the migration, but it's a sufficient start to witness the value.

A good follow-up would be to adopt TaskCenter's potential of scheduling tasks on multiple runtimes and move the rocksdb
single-threaded writer runtime to be managed by TaskCenter as well, but that's for another time.