Send ingress responses directly to ingress
This commit changes how ingress responses are sent to the ingress.
Before, they went through the outbox. Now, they are directly sent
to the ingress. This extends a bit the cases in which we cannot deliver
an ingress response to the ingress but this does not matter too much
since we cannot guarantee ingress response delivery in the general
case.

As part of this change, this commit also removes no longer used parts
of the network router and the ingress acking.

This fixes #1210.