move restate-task-center to restate-core
We need to consolidate a few bits of the system into a core crate to avoid the unnecessary dependency management hassle.

The end goal is to consolidate many of the shared components into this crate. At the moment, task-center and metadata appear to be the best candidate since they are used from everywhere.