Remove batching from PartitionProcessor
Since we are no longer requiring to flush writes to RocksDB, the existing
batching mechanism might no longer needed. Moreover, it adds complexity which
could be avoided until we assert that batching of commands has a measurable
impact on Restate's performance. Until then and in order to make the future
development a bit easier, we remove it.