Integrate local metadata store into the Node
This commit integrates the local metadata store into the Node.
It achieves it by introducing the MetadataStore role and requiring
that the bootstrap node runs it. The role is the first started role.
It runs the LocalMetadataStoreService. After starting the MetadataStore,
the node tries to upsert the new NodeConfig with which it announces its
NodeId, address and roles it runs.

This fixes #1296.