Add cluster marker and validate it at start up of Node
This commit introduces the ClusterMarker which is a Json encoded
file containing information about the last user of Restate's data
directory. That way, we can validate that only a compatible Restate
version and a process belonging to the right cluster opens the data
directory. The cluster marker is stored in the file .cluster-marker.

This fixes #1391.