Partition-scoped storage reader in invoker
This changes the invoker to allow partitions to pass down scoped storage readers along registration messages in invoker. This means that the invoker uses the partition-supplied storage reader for invocations owned by this particular partition.
Additionally, this attempts to consolidate (in most places) the journal reader and state reader under a single generic type.