Betters shutdown logging
When using `#[error(transparent)]` to wrap `ShutdownError`, the error root cause is not propagated correctly all the way down to task center. The reason is that `thiserror` will forward `source()` to the type marked with `transparent`. In this case, it's ShutdownError it self (and it's not a root cause for itself). The solution is to make a private marker type that's guaranteed to always be the root cause of an error chain that was caused by a shutdown.