Prevent race condition in ResponseTracker by using entry which holds the lock
This commit resolves the race condition in the ResponseTracer when creating a new
token with an existing correlation id and completing the very same correlation id
at the same time. The solution is to use the entry method of Dashmap which keeps the
lock as long as entry is not dropped.

This fixes #1531.