[introspection] Set the minimal query parallelism to 2
This commit makes sure that the default parallelism for the query engine
(datafusion) is at least 2. This value translates to 'target partitions'
which is used by datafusion during query execution to control execution
concurrency. Setting this value to 1 however causes #1585, and setting
this to high is unadvised in our case, as datafusion is used for
low-prio introspection tasks.
However once issue #1585 is resolved, we will return to properly sizing
datafusion.