Introduce new state machine infra and related test infra.
The new infra uses `StateMachineContext` as entrypoint, and from there every command struct implements `ApplicableCommand`. `ApplicableCommand` applies the command by using the `CommandContext`, which contains storage and actions collector.

To facilitate the transition, I've introduced a MaybeApplicableCommand which is implemented for `Command`, and returns the command in case the new infra doesn't yet support that specific command.