Let shuffle always use get_next_message to support holey outboxes
This commit changes the shuffle to always use get_next_message to read
the next outbox message. Before we were using get_message which was only
looking at a specific outbox entry. If this outbox entry was empty, then
the shuffle assumed that the outbox is empty. This did not work if the
outbox contained holes. Now with get_next_message, we always scan until
the next outbox message.

The change itself is trivial. In order to ensure that the change works,
this commit added a few unit tests to ensure the desired behaviour.

Note: With replacing get_message with get_next_message, the shuffle will
perform for every read a RocksDB scan operation. This is highly inefficient
and we should replace this logic with a tailing iterator.

This fixes #1639.