Fix errornous enable of test-util
default members of the workspace need to avoid enabling `test-util` feature. Feature unification ends enabling restate-core test-util in restate-server which is not what we want.
I've added a message to task-center to make it visible when this happens, but this can also happen with restate-types which comes with more severe consequences (base-dir being in /tmp by default).

This PR addresses the immediate issue by moving restate-benchmarks out of crates and making sure that default-members don't have any crate that enables `test-util` transitively.