Support http1 for request-response protocol deployments (#1232)
* Support http1 for request-response protocol deployments

We already have the plumbing to set the request version appropriately based on discovered version (discovery goes through HTTP1.1 by default). We just end up ignoring that version at the last minute by setting the client to be http2 only. With this change, we will discover with http2, and fallback to http1.1 if it doesn't appear to be supported. The fallback request will respect ALPN, so it could technically still end up using http2 against HTTPS endpoints, but this would be odd.

* Accept a flag to force the use of http2

* Add flag to CLI

* Add comments