Only emit admin API incompatibility warning if cluster is running
This commit changes when the CLI prints an admin API incompatibility
warning. The warning is now printed if the admin API version couldn't
be validated and if the cluster is reachable (via checking the /health
endpoint). If the cluster is not reachable, then the CLI terminates with
an error message stating that the Restate server is not reachable.

This fixes #1646.