Remove two default imports from datafusion dependency (#1679)
Primary goal is to remove 'compression' which bundles in various
compression libs, one of which appears to be dynamically linked for mac
in some cases. We don't need the compression feature

As a extra benefit, we can remove parquet support in the interest of
compile time and binary size. I don't really see how this can be used
usefully as we have disabled functions that read and write files.