Request metadata updates from other nodes
This commit updates the network reactor to check the header of
the remote node for newer metadata versions. If the remote node
should have newer metadata, the reactor informs the MetadataFetcher
which tries to fetch the metadata updates from the remote node.
The MetadataFetcher makes sure that requests are regularly sent in
case that messages are being lost. Moreover, it ensures that only
one node that announced a given metadata version will be asked for
the update.

This fixes #1697.